Added calc of area, centroid and bbox for map obj
Now the area and centroid is calculated for each area of each country.
Then a bounding box is defined as the bounding box around the largest
area. I also determine -- for each area -- if it may be in need of special
highlighting to easier spot it. The calculation takes a bit of time but is
acceptable.